OPENAI_IMAGE_GENERATION/requirements.txt
requests==2.31.0
openai == 1.2.3
python-dotenv==1.0.0
Pillow==10.1.0